refactor: Remove low-level artifact wrapper types, unify into artifact.rs
Remove ArtifactArchive, ArtifactDir, ArtifactArchiveBuilder,
ArtifactDirBuilder pyclass types that were only used as internal
implementation details. Merge free functions into artifact.rs
and delete builder.rs.
