--- docs/api_reference/api/_items/ommx.artifact.ArtifactArchive.rst
@@ -1,4 +1,0 @@
-ArtifactArchive
-===============
-
-.. pyo3-api-class:: ommx.artifact ArtifactArchive
