--- docs/api_reference/api/_items/ommx.artifact.ArtifactArchiveBuilder.rst
@@ -1,4 +1,0 @@
-ArtifactArchiveBuilder
-======================
-
-.. pyo3-api-class:: ommx.artifact ArtifactArchiveBuilder
