--- docs/api_reference/api/_items/ommx.artifact.ArtifactDir.rst
@@ -1,4 +1,0 @@
-ArtifactDir
-===========
-
-.. pyo3-api-class:: ommx.artifact ArtifactDir
