--- docs/api_reference/api/_items/ommx.artifact.ArtifactDirBuilder.rst
@@ -1,4 +1,0 @@
-ArtifactDirBuilder
-==================
-
-.. pyo3-api-class:: ommx.artifact ArtifactDirBuilder
